Bug fixes and added new shuffle type

+++ app/src/main/java/com/example/musicplayer/SongListAdapter.java
@@ -3,6 +3,7 @@ package com.example.musicplayer;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.BitmapFactory;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -46,13 +47,21 @@ public class SongListAdapter extends RecyclerView.Adapter<SongListAdapter.ViewHo
             @Override
             public void onClick(View view) {
                 //Navigate to another activity
-                MyMusicPlayer.getInstance().reset();
-                MyMusicPlayer.currentIndex = position;
-                Intent intent = new Intent(context, MusicPlayerActivity.class);
-                intent.putExtra("LIST",songsList);
-                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                context.startActivity(intent);
+                int currentPosition = holder.getAdapterPosition(); // Always get the current position
+                if (currentPosition != RecyclerView.NO_POSITION) {
+                    MyMusicPlayer.currentIndex = currentPosition;
 
+                    Intent intent = new Intent(context, MusicPlayerActivity.class);
+
+                    ArrayList<Integer> songIdList = new ArrayList<Integer>();
+                    for (Song song : songsList) {
+                        songIdList.add(song.getSongId());
+                    }
+
+                    intent.putExtra("LIST", songIdList);
+                    intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                    context.startActivity(intent);
+                }
             }
         });
 
@@ -60,23 +69,25 @@ public class SongListAdapter extends RecyclerView.Adapter<SongListAdapter.ViewHo
         holder.itemView.setOnLongClickListener(new View.OnLongClickListener() {
             @Override
             public boolean onLongClick(View view) {
-                PopupMenu popupMenu = new PopupMenu(context, view);
-                popupMenu.getMenuInflater().inflate(R.menu.overflow_menu_item, popupMenu.getMenu());
+                int currentPosition = holder.getAdapterPosition(); // Always get the current position
+                if (currentPosition != RecyclerView.NO_POSITION) {
+                    PopupMenu popupMenu = new PopupMenu(context, view);
+                    popupMenu.getMenuInflater().inflate(R.menu.overflow_menu_item, popupMenu.getMenu());
 
-                // Set menu item click listener
-                popupMenu.setOnMenuItemClickListener(menuItem -> {
-                        if(menuItem.getItemId() == R.id.delete_item) {
+                    // Set menu item click listener
+                    popupMenu.setOnMenuItemClickListener(menuItem -> {
+                        if (menuItem.getItemId() == R.id.delete_item) {
                             if (context instanceof LibraryActivity) {
                                 ((LibraryActivity) context).deleteSong(songData);
-                            }
-                            else if (context instanceof PlayListActivity) {
+                            } else if (context instanceof PlayListActivity) {
                                 ((PlayListActivity) context).deleteSong(songData);
                             }
                         }
-                        return false;
-                });
-                popupMenu.show();
-                return false;
+                        return true;
+                    });
+                    popupMenu.show();
+                }
+                return true;
             }
         });
     }
